feat: add rudimentary CLI --watch mode
