+++ src/cli/options.ts
@@ -1,0 +1,20 @@
+import { parseArgs, ParseArgsOptionsConfig } from "node:util";
+
+export const options = {
+	fix: {
+		type: "boolean",
+	},
+	help: {
+		type: "boolean",
+	},
+	version: {
+		type: "boolean",
+	},
+	watch: {
+		type: "boolean",
+	},
+} satisfies ParseArgsOptionsConfig;
+
+export type OptionsValues = ReturnType<
+	typeof parseArgs<{ options: typeof options }>
+>["values"];
