+++ src/cli/runCli.ts
@@ -1,31 +1,14 @@
-import { debugForFile } from "debug-for-file";
-import path from "node:path";
 import { parseArgs } from "node:util";
 
-import { writeToCache } from "../cache/writeToCache.js";
-import { isConfig } from "../configs/isConfig.js";
-import { runPrettier } from "../formatting/runPrettier.js";
-import { plainReporter } from "../reporters/plain.js";
-import { runConfig } from "../running/runConfig.js";
-import { runConfigFixing } from "../running/runConfigFixing.js";
-import { findConfigFileName } from "./findConfigFileName.js";
+import { plainPresenter } from "../presenters/plain.js";
+import { options } from "./options.js";
 import { packageData } from "./packageData.js";
-
-const log = debugForFile(import.meta.filename);
+import { runCliSingleRun } from "./runCliSingleRun.js";
+import { runCliWatch } from "./runCliWatch.js";
 
 export async function runCli() {
 	const { values } = parseArgs({
-		options: {
-			fix: {
-				type: "boolean",
-			},
-			help: {
-				type: "boolean",
-			},
-			version: {
-				type: "boolean",
-			},
-		},
+		options,
 		strict: true,
 	});
 
@@ -43,57 +26,7 @@ export async function runCli() {
 		return 0;
 	}
 
-	const configFileName = await findConfigFileName(process.cwd());
-	if (!configFileName) {
-		console.error("No flint.config.* file found");
-		return 2;
-	}
-
-	const { default: config } = (await import(
-		path.join(process.cwd(), configFileName)
-	)) as {
-		default: unknown;
-	};
-
-	if (!isConfig(config)) {
-		console.error(
-			`${configFileName} does not default export a Flint defineConfig value.`,
-		);
-		return 2;
-	}
-
-	log("Running with Flint config: %s", configFileName);
-
-	const configDefinition = {
-		...config.definition,
-		filePath: configFileName,
-	};
-
-	const configResults = await (values.fix
-		? runConfigFixing(configDefinition)
-		: runConfig(configDefinition));
-
-	// TODO: Eventually, it'd be nice to move everything fully in-memory.
-	// This would be better for performance to avoid excess file system I/O.
-	// https://github.com/JoshuaKGoldberg/flint/issues/73
-	const [formattingResults] = await Promise.all([
-		runPrettier(configResults.allFilePaths, values.fix),
-		writeToCache(configFileName, configResults),
-	]);
-
-	for (const line of plainReporter(configResults, formattingResults)) {
-		console.log(line);
-	}
-
-	if (formattingResults.dirty.size && !formattingResults.written) {
-		return 1;
-	}
-
-	for (const fileResults of configResults.filesResults.values()) {
-		if (fileResults.reports.length) {
-			return 1;
-		}
-	}
-
-	return 0;
+	return await (values.watch
+		? runCliWatch(values)
+		: runCliSingleRun(values, plainPresenter));
 }
